fix event dispatch

+++ Code/encoding/src/Messages/ServerTimeSettings.cpp
@@ -4,8 +4,8 @@
 void ServerTimeSettings::SerializeRaw(TiltedPhoques::Buffer::Writer &aWriter) const noexcept
 {
     Serialization::WriteVarInt(aWriter, ServerTick);
-    Serialization::WriteVarInt(aWriter, static_cast<uint64_t>(TimeScale));
-    Serialization::WriteVarInt(aWriter, static_cast<uint64_t>(Time));
+    Serialization::WriteVarInt(aWriter, static_cast<uint32_t>(TimeScale));
+    Serialization::WriteVarInt(aWriter, static_cast<uint32_t>(Time));
 }
 
 void ServerTimeSettings::DeserializeRaw(TiltedPhoques::Buffer::Reader &aReader) noexcept
